+++ data/src/main/java/com/sopt/now/data/repositoryimpl/UserInfoRepositoryImpl.kt
@@ -1,7 +1,7 @@
-package com.sopt.now.repositoryimpl
+package com.sopt.now.data.repositoryimpl
 
-import com.sopt.now.datasource.SharedPreferenceDataSource
-import com.sopt.now.dto.UserDto
+import com.sopt.now.data.datasource.SharedPreferenceDataSource
+import com.sopt.now.data.dto.UserDto
 import com.sopt.now.domain.entity.UserEntity
 import com.sopt.now.domain.repository.UserInfoRepository
 import javax.inject.Inject
